refactor(component): build PlayerPage component to reuse it
